delete get checks on controller

+++ src/Services/Catalog/Catalog.API/Infrastructure/CatalogRepository.cs
@@ -12,29 +12,49 @@ public class CatalogRepository : ICatalogRepository
         _context = context ?? throw new ArgumentNullException(nameof(context));
     }
 
-    public async Task SaveChangesAsync()
-    {
-        await _context.SaveChangesAsync();
-    }
-
-    public async Task<CatalogItem> CreateAsync(CatalogItem item)
+    public async Task<CatalogItem?> CreateAsync(CatalogItem item)
     {
         var createdItem = await _context.CatalogItems.AddAsync(item);
 
-        return createdItem.Entity;
+        return createdItem?.Entity ?? null;
     }
 
-    public void Update(CatalogItem item)
+    public async Task<CatalogItem?> UpdateAsync(CatalogItem item)
     {
+        if (item is null)
+        {
+            return null;
+        }
+
+        var catalogItem = await _context.CatalogItems.FindAsync(item.Id);
+
+        if (catalogItem is null)
+        {
+            return null;
+        }
+
         _context.CatalogItems.Update(item);
+        await _context.SaveChangesAsync();
+
+        return item;
     }
 
-    public void Delete(CatalogItem item)
+    public async Task<CatalogItem?> DeleteAsync(long id)
     {
+        var item = await _context.CatalogItems.FindAsync(id);
+
+        if (item is null)
+        {
+            return null;
+        }
+
         _context.CatalogItems.Remove(item);
+        await _context.SaveChangesAsync();
+
+        return item;
     }
 
-    public async Task<CatalogItem> GetAsync(long id, bool asNoTracking = false)
+    public async Task<CatalogItem?> GetAsync(long id, bool asNoTracking = false)
     {
         return asNoTracking ?
             await _context.CatalogItems.AsNoTracking().SingleOrDefaultAsync(c => c.Id == id)
